New ticket model refactor
* Single table inheritance to mixins
* Remove quite a lot of fields which will need wiring back up
